+++ web/wiki/src/components/molecules/Notifications/index.ts
@@ -1,3 +1,0 @@
-export { Notifications } from './Notifications'
-export { notificationsReducer } from './reducer'
-export { NotificationsProvider } from './context'