+++ web/wiki/src/components/organisms/CardResourceLink.tsx
@@ -1,7 +1,6 @@
 import styled from 'styled-components'
-import { FlexBox, dimensions } from '@itacademy/ui'
-import { ResourceTitleLink } from '../molecules'
-import EditResource from './EditResource'
+import { FlexBox, ResourceTitleLink, dimensions } from '@itacademy/ui'
+import { EditResource } from './EditResource'
 import { TCardResource } from '../../types'
 
 const CardContainerStyled = styled(FlexBox).withConfig({